Relocate/rename ode files
