--- inst/include/plant/individual_tools.h
@@ -1,6 +1,0 @@
-// -*-c++-*-
-
-#ifndef INDIVIDUAL_TOOLS
-#define INDIVIDUAL_TOOLS_
-
-#endif /* INDIVIDUAL_TOOLS */
