+++ inst/include/plant/node.h
@@ -4,7 +4,7 @@
 
 #include <plant/environment.h>
 #include <plant/gradient.h>
-#include <plant/ode_interface.h>
+#include <plant/ode_solver/ode_interface.h>
 
 namespace plant {
